Move server pkg
